+++ client/player_look.cc
@@ -10,21 +10,21 @@
 #include <shared/head.hh>
 #include <shared/world.hh>
 
-constexpr static const float PITCH_MIN = -1.0f * cxmath::radians(89.9f);
-constexpr static const float PITCH_MAX = +1.0f * cxmath::radians(89.9f);
+constexpr static const double PITCH_MIN = -1.0 * cxmath::radians(89.9);
+constexpr static const double PITCH_MAX = +1.0 * cxmath::radians(89.9);
 
-static float previous_cx = 0.0f;
-static float previous_cy = 0.0f;
+static double previous_cx = 0.0;
+static double previous_cy = 0.0;
 
 static void on_cursor_pos(const CursorPosEvent &event)
 {
     if(globals::world.registry.valid(globals::player)) {
-        const float dx = event.xpos - previous_cx;
-        const float dy = event.ypos - previous_cy;
+        const double dx = event.xpos - previous_cx;
+        const double dy = event.ypos - previous_cy;
 
         auto &head = globals::world.registry.get<HeadComponent>(globals::player);
-        head.angles.x -= cxmath::radians(dy) * 0.25f; // UNDONE
-        head.angles.y -= cxmath::radians(dx) * 0.25f;
+        head.angles.x -= cxmath::radians(dy) * 0.25; // UNDONE
+        head.angles.y -= cxmath::radians(dx) * 0.25;
         head.angles.x = cxmath::clamp(head.angles.x, PITCH_MIN, PITCH_MAX);
         head.angles = angle::wrap_180_n(head.angles);
     }
@@ -35,11 +35,11 @@ static void on_cursor_pos(const CursorPosEvent &event)
 
 void player_look::init()
 {
-    float width, height;
+    double width, height;
     screen::get_size(width, height);
 
-    previous_cx = width / 2.0f;
-    previous_cy = height / 2.0f;
+    previous_cx = width / 2.0;
+    previous_cy = height / 2.0;
 
     globals::dispatcher.sink<CursorPosEvent>().connect<&on_cursor_pos>();
 }
